lemonade-release-2023.6.1.xml: fix host address
On Commit 74cedfb725de18ead8fd8cb030a5e5c99ec82505 ("vcpi: redo work
which was not covered" / Silvio Fricke <silvio.fricke@softing.com> / Mon
Jun 5 10:18:29 2023 +0200) we changed the host adress to a `.` which
reults in a pull from the right repository folder.

The subsequent repository pull is executed differently with 2 points
than with one. Here is an example:

With two '..' dots:

    $ repo sync
    [...]
    error: Cannot fetch VCPI.android.packages.modules.Connectivity from https://dev.azure.com/SoftingAutomotiveElectronics/Automotive/VCPI.android.packages.modules.Connectivity

The right url is
https://dev.azure.com/SoftingAutomotiveElectronics/Automotive/_git/VCPI.android.packages.modules.Connectivity
(with _git in path).

Change-Id: I255a935593b81de43eee86ca36f6e21e481a1d46
Signed-off-by: Silvio Fricke <silvio.fricke@softing.com>