lemonade: add develop approach (WIP)
These files has different purposes.

The devel-file is including softing-all and default-fixed file. This
file containes branch information and not commit ids. softing-all is a
sepcial file, it will be used by every phone therefore we have to track
the information in a own file.

default-fixed containes commit numbers of all other software we use on
our porduct which are not changed by softing.

Change-Id: Ic683cb59cd4fdb522b4869d02c502b186526f140
Signed-off-by: Silvio Fricke <silvio.fricke@softing.com>