az: android-build-all-phones: refractor build steps to template file and remove matrix strategy
The main problem on a matrix strategy is, we cannot use demand
information in selection of the capabilities of a agent.

I have now removed the matrix strategy and because I don't want to write
all the stuff two times (one per phone) I moved the whole stuff to a
template file.

Change-Id: Ia06b2420abe2291aa31731a131227ec320b2a11b
Signed-off-by: Silvio Fricke <silvio.fricke@softing.com>